lessvote query fixed

+++ CodeFirstDB/ViewModle/Vote.cs
@@ -28,12 +28,12 @@ namespace CodeFirstDB.ViewModle
         [ForeignKey("UserID")]
         public int UserId { get; set; }
         public virtual User User { get; set; }
-        
+
         //public int UserId { get; set; }
 
         [Column("Value")]
         [Required]
-        public int Value { get; set; }
+        public long Value { get; set; }
     }
 
 }
